Adición de campos a tablas de carreras, ciclos, planes, documentos y evaluaciones

+++ database/migrations/2019_10_21_213506_create_cycles_table.php
@@ -15,7 +15,7 @@ class CreateCyclesTable extends Migration
     {
         Schema::create('cycles', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedBigInteger('level_id')->unsigned();
+            $table->BigInteger('level_id')->unsigned();
             $table->foreign('level_id')->references('id')->on('levels');
             $table->dateTime('start_date');
             $table->dateTime('end_date');
